refactor: ExternalLinkコンポーネントに切り出し
